procstat: add support for histogram stats on series.
This statistics allows to perform analisys on series of integral data points.
We currenly allow to register up to 20 different percentile points of interest.
Procstat will create readable file for every "percentile". When points are added to the series
histogram is updated, and when appropriate file is read, percentile is calculated.