histogram: allow user to control percentile canculation.
When percentiles are calculated heavy work performed, because we have to go
over the whole histogram. Eventhough this is not on hot path, this clculation still consumes lots of resources.
Moreover when whole directory is read (via grep -r '' for example) whole percentiles will be recalculated for every file
This change allows user to control percentiles calculation, for example only recalculating it after specific time.