aggregator: only generate full lines
Running out of the output buffer space was handled in a naive way.
We supplied the entire output to the caller and recorded the
number of characters to skip from the current item output
in the following call.

This works well when the VALUES of items do not change between calls,
so the same set of items generates the same number of output characters.
This however is not the case: values change, so previous 9 changing to 10
would produce a 1-char shift. This may result in garbled output.

Instead we now count output lines, record how many have been output
in the current root item, and skip this number of lines the next time.
When the output buffer becomes full the last incomplete line is not
generated. Instead we pad spaces at the end of the previous line.
This way we still return the exact number of bytes requested,
so the caller may not detect end of file.

Issue: LBM1-15406

Signed-off-by: Anton Eidelman <anton@lightbitslabs.com>