procstat: use array of mutex to lock/unlock subtrees.
Associate a mutex identified by an 8-bit index with the root node and
all immediate subdirectories of the root.

Add the "nearest active" mutex parameter to the recursive functions
item_put_locked() / item_put_children_locked().  Ensure that the
appropriate lock is in place before descending a directory but avoid
double-locking.

Issue: LBM1-19873